requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7

